Make snakeyaml-engine and jackson-databind optional at runtime
Split DeclarativeConfiguration into two classes:
- DeclarativeConfiguration: create(model) only, no YAML/jackson-databind deps
- DeclarativeConfigurationParser: all parsing (MAPPER static block, parse,
  loadYaml, parseAndCreate, toConfigProperties, createSampler, env-var
  substitution, snakeyaml inner classes)

Change snakeyaml-engine and jackson-databind from implementation to
compileOnly in build.gradle.kts (keeping jackson-annotations as api since
the generated model POJOs use @JsonProperty etc. as public API).
Remove the unused jackson-dataformat-yaml dependency.

Users who only call DeclarativeConfiguration.create(model) for programmatic
SDK configuration no longer incur transitive YAML/jackson-databind deps.
Users who need YAML parsing add snakeyaml-engine and jackson-databind and
call DeclarativeConfigurationParser.parseAndCreate(stream).

This also opens the door for merging declarative config into autoconfigure,
since the static initializer that caused NoClassDefFoundError when
jackson-databind was absent is now isolated in DeclarativeConfigurationParser.
