+++ api/incubator/src/test/java/io/opentelemetry/api/incubator/ExtendedOpenTelemetryTest.java
@@ -25,7 +25,7 @@ import io.opentelemetry.api.testing.internal.AbstractOpenTelemetryTest;
 import io.opentelemetry.api.trace.TracerProvider;
 import io.opentelemetry.context.propagation.ContextPropagators;
 import io.opentelemetry.sdk.OpenTelemetrySdk;
-import io.opentelemetry.sdk.extension.incubator.fileconfig.DeclarativeConfiguration;
+import io.opentelemetry.sdk.extension.incubator.fileconfig.DeclarativeConfigurationParser;
 import io.opentelemetry.sdk.internal.ExtendedOpenTelemetrySdk;
 import io.opentelemetry.sdk.internal.SdkConfigProvider;
 import java.io.ByteArrayInputStream;
@@ -94,7 +94,7 @@ class ExtendedOpenTelemetryTest extends AbstractOpenTelemetryTest {
 
     SdkConfigProvider configProvider =
         SdkConfigProvider.create(
-            DeclarativeConfiguration.toConfigProperties(
+            DeclarativeConfigurationParser.toConfigProperties(
                 new ByteArrayInputStream(configYaml.getBytes(StandardCharsets.UTF_8))));
     ExtendedOpenTelemetry openTelemetry =
         ExtendedOpenTelemetrySdk.create(OpenTelemetrySdk.builder().build(), configProvider);
