+++ api/incubator/src/test/java/io/opentelemetry/api/incubator/config/InstrumentationConfigUtilTest.java
@@ -8,7 +8,7 @@ package io.opentelemetry.api.incubator.config;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import com.google.common.collect.ImmutableMap;
-import io.opentelemetry.sdk.extension.incubator.fileconfig.DeclarativeConfiguration;
+import io.opentelemetry.sdk.extension.incubator.fileconfig.DeclarativeConfigurationParser;
 import io.opentelemetry.sdk.extension.incubator.fileconfig.YamlDeclarativeConfigProperties;
 import io.opentelemetry.sdk.internal.SdkConfigProvider;
 import java.io.ByteArrayInputStream;
@@ -61,7 +61,7 @@ class InstrumentationConfigUtilTest {
 
   private static ConfigProvider toConfigProvider(String configYaml) {
     return SdkConfigProvider.create(
-        DeclarativeConfiguration.toConfigProperties(
+        DeclarativeConfigurationParser.toConfigProperties(
             new ByteArrayInputStream(configYaml.getBytes(StandardCharsets.UTF_8))));
   }
