+++ sdk-extensions/incubator/src/main/java/io/opentelemetry/sdk/extension/incubator/fileconfig/OpenTelemetryConfigurationFactory.java
@@ -43,7 +43,7 @@ final class OpenTelemetryConfigurationFactory
   public DeclarativeConfigResult create(
       OpenTelemetryConfigurationModel model, DeclarativeConfigContext context) {
     DeclarativeConfigProperties modelProperties =
-        DeclarativeConfiguration.toConfigProperties(model, context.getDelegateComponentLoader());
+        DeclarativeConfigurationParser.toConfigProperties(model, context.getDelegateComponentLoader());
     SdkConfigProvider sdkConfigProvider = SdkConfigProvider.create(modelProperties);
     context.setConfigProvider(sdkConfigProvider);
     OpenTelemetrySdkBuilder builder =
